Added Open Source Cards in React Section
