+++ src/pages/Experience/top_ilus.js
@@ -1,0 +1,115 @@
+import React from 'react';
+
+function topIlus() {
+
+    return (
+        <div>
+        <div class="container pt-5" style={{width: "100%", minHeight: "100%"}}>
+            <section>
+            <div class="container-fluid px-3">
+                <div class="row">
+                    <div class="godown col-md-6">
+                        <svg viewBox="0 0 320 266" fill="none" xmlns="http://www.w3.org/2000/svg" width="90%">
+            <g id="shadows">
+                <path id="Vector" opacity="0.5" d="M185.642 35.2437C165.167 34.5318 145.665 27.9228 127.044 20.6905C108.422 13.4582 89.9772 5.44444 69.9919 1.61315C57.1376 -0.8514 42.4358 -1.19941 32.0808 5.69121C22.1119 12.3351 18.8912 23.772 17.1607 34.3958C15.8572 42.3874 15.0916 50.7998 18.6603 58.2788C21.1406 63.4737 25.5382 67.8397 28.5818 72.8162C39.1708 90.1282 31.6854 111.477 20.2137 128.381C14.8353 136.312 8.58694 143.883 4.43295 152.318C0.278964 160.752 -1.64142 170.43 1.99371 179.042C5.59404 187.584 14.1773 193.987 23.4755 198.496C42.3598 207.655 64.6041 210.274 86.3105 211.758C134.345 215.045 182.634 213.622 230.795 212.198C248.62 211.669 266.52 211.135 284.053 208.379C293.791 206.848 303.843 204.421 310.914 198.556C319.886 191.112 322.11 178.507 316.096 169.171C306.01 153.514 278.131 149.625 271.073 132.823C267.194 123.572 271.177 113.271 276.818 104.691C288.916 86.2906 309.193 70.146 310.259 49.1135C310.993 34.6679 301.245 20.2001 286.173 13.3633C270.377 6.20057 248.468 7.09907 236.819 18.9599C224.813 31.1719 203.723 35.8701 185.642 35.2437Z" fill="#bb86fc"></path>
+                <path id="Vector_2" opacity="0.5" d="M153.037 260.094C206.15 260.094 249.206 254.27 249.206 247.085C249.206 239.9 206.15 234.075 153.037 234.075C99.9253 234.075 56.8694 239.9 56.8694 247.085C56.8694 254.27 99.9253 260.094 153.037 260.094Z" fill="#bb86fc"></path>
+                <path id="Vector_3" opacity="0.1" d="M273.475 246.104C278.18 246.104 281.995 245.46 281.995 244.664C281.995 243.869 278.18 243.225 273.475 243.225C268.769 243.225 264.955 243.869 264.955 244.664C264.955 245.46 268.769 246.104 273.475 246.104Z" fill="#bb86fc"></path>
+                <path id="Vector_4" opacity="0.1" d="M28.8925 257.917C33.5979 257.917 37.4124 257.273 37.4124 256.478C37.4124 255.683 33.5979 255.038 28.8925 255.038C24.1871 255.038 20.3726 255.683 20.3726 256.478C20.3726 257.273 24.1871 257.917 28.8925 257.917Z" fill="#bb86fc"></path>
+                <path id="Vector_5" opacity="0.1" d="M61.849 265.64C66.5545 265.64 70.369 264.996 70.369 264.201C70.369 263.405 66.5545 262.761 61.849 262.761C57.1436 262.761 53.3291 263.405 53.3291 264.201C53.3291 264.996 57.1436 265.64 61.849 265.64Z" fill="#bb86fc"></path>
+            </g>
+
+            <g id="Phone">
+                <path id="Vector_42" d="M205.518 91.6914H204.163C203.763 91.6914 203.439 92.0158 203.439 92.4159V114.069C203.439 114.469 203.763 114.793 204.163 114.793H205.518C205.918 114.793 206.242 114.469 206.242 114.069V92.4159C206.242 92.0158 205.918 91.6914 205.518 91.6914Z" fill="#3A3768"></path>
+                <path id="Vector_43" d="M111.137 73.5917H111.014C110.613 73.5917 110.289 73.916 110.289 74.3162V80.4728C110.289 80.8729 110.613 81.1973 111.014 81.1973H111.137C111.537 81.1973 111.861 80.8729 111.861 80.4728V74.3162C111.861 73.916 111.537 73.5917 111.137 73.5917Z" fill="#3A3768"></path>
+                <path id="Vector_44" d="M111.279 87.5184H110.957C110.557 87.5184 110.232 87.8428 110.232 88.2429V100.031C110.232 100.431 110.557 100.756 110.957 100.756H111.279C111.68 100.756 112.004 100.431 112.004 100.031V88.2429C112.004 87.8428 111.68 87.5184 111.279 87.5184Z" fill="#3A3768"></path>
+                <path id="Vector_45" d="M111.223 105.387H110.985C110.585 105.387 110.261 105.712 110.261 106.112V118.014C110.261 118.414 110.585 118.738 110.985 118.738H111.223C111.623 118.738 111.947 118.414 111.947 118.014V106.112C111.947 105.712 111.623 105.387 111.223 105.387Z" fill="#3A3768"></path>
+                <path id="Vector_46" d="M192.616 48.459H123.514C116.701 48.459 111.178 53.9817 111.178 60.7944V226.704C111.178 233.517 116.701 239.039 123.514 239.039H192.616C199.429 239.039 204.951 233.517 204.951 226.704V60.7944C204.951 53.9817 199.429 48.459 192.616 48.459Z" fill="#3f3d56"></path>
+                <path id="Vector_47" d="M191.724 53.4102H180.967V54.7042C180.967 55.5096 180.808 56.3072 180.5 57.0513C180.192 57.7954 179.74 58.4714 179.17 59.0408C178.6 59.6102 177.924 60.0618 177.18 60.3697C176.436 60.6776 175.638 60.8359 174.833 60.8355H140.623C139.818 60.8359 139.02 60.6776 138.276 60.3697C137.531 60.0618 136.855 59.6102 136.285 59.0408C135.716 58.4714 135.264 57.7954 134.956 57.0513C134.647 56.3072 134.489 55.5096 134.489 54.7042V53.4102H124.418C122.443 53.4102 120.549 54.1949 119.152 55.5915C117.756 56.9882 116.971 58.8825 116.971 60.8577V226.628C116.971 228.603 117.756 230.497 119.152 231.894C120.549 233.291 122.443 234.075 124.418 234.075H191.724C193.699 234.075 195.593 233.291 196.99 231.894C198.387 230.497 199.171 228.603 199.171 226.628V60.8703C199.173 59.8912 198.981 58.9214 198.608 58.0164C198.234 57.1114 197.686 56.2889 196.994 55.596C196.303 54.9031 195.481 54.3534 194.577 53.9783C193.672 53.6033 192.703 53.4102 191.724 53.4102Z" fill="#bb86fc"></path>
+                <path id="Vector_48" d="M160.89 54.1505H150.244C149.568 54.1505 149.02 54.6987 149.02 55.3749V55.5837C149.02 56.2599 149.568 56.808 150.244 56.808H160.89C161.566 56.808 162.114 56.2599 162.114 55.5837V55.3749C162.114 54.6987 161.566 54.1505 160.89 54.1505Z" fill="#bb86fc"></path>
+                <path id="Vector_49" d="M165.604 56.9884C166.437 56.9884 167.113 56.3127 167.113 55.4793C167.113 54.6458 166.437 53.9702 165.604 53.9702C164.77 53.9702 164.095 54.6458 164.095 55.4793C164.095 56.3127 164.77 56.9884 165.604 56.9884Z" fill="#bb86fc"></path>
+                <path id="Vector_50" opacity="0.1" d="M190.803 71.5574H125.33V210.521H190.803V71.5574Z" fill="black"></path>
+            </g>
+            <g id="Man">
+                <path id="Vector_51" d="M188.797 249.315C188.797 249.315 187.253 245.721 186.564 244.063C186.175 243.114 185.516 242.956 185.013 242.978L185.042 242.662L185.165 241.165L185.191 240.849C186.383 240.599 187.405 240.403 187.405 240.403L186.662 202.02C186.662 202.02 186.345 201.615 185.58 199.173C185.2 197.955 183.365 196.459 181.65 195.266C181.381 195.079 181.119 194.902 180.859 194.731C181.484 194.208 182.14 193.724 182.824 193.282C183.799 192.716 184.121 189.486 184.121 189.486L183.96 189.802C184.121 189.486 183.96 181.238 183.96 181.238L183.39 172.677C183.39 172.677 182.985 162.499 179.822 159.832C176.658 157.165 172.662 156.232 172.662 156.232L170.448 155.662L166.366 154.615C166.252 154.428 166.142 154.232 166.05 154.027C166.009 153.947 165.971 153.865 165.933 153.783C167.18 152.9 168.197 151.731 168.899 150.374C169.601 149.017 169.967 147.512 169.967 145.984C169.967 145.471 169.926 144.959 169.843 144.453C169.843 144.358 169.809 144.26 169.79 144.165C170.125 143.936 170.442 143.681 170.739 143.403C171.222 143.018 171.617 142.533 171.897 141.982C171.94 141.874 171.976 141.763 172.004 141.65C172.004 141.618 172.004 141.587 172.023 141.552C172.042 141.517 172.052 141.384 172.061 141.299C172.061 141.267 172.061 141.236 172.061 141.204C172.061 141.112 172.061 141.017 172.061 140.925C172.061 140.903 172.061 140.881 172.061 140.856C172.061 140.751 172.045 140.647 172.029 140.54C172.031 140.533 172.031 140.527 172.029 140.521C171.972 140.153 171.876 139.793 171.745 139.445C171.428 138.629 171.017 137.863 170.868 136.993C170.685 135.965 170.897 134.842 170.406 133.918C170.203 133.569 169.925 133.269 169.591 133.039C169.258 132.81 168.879 132.658 168.48 132.592C167.677 132.468 166.855 132.562 166.101 132.864C165.732 133.04 165.334 133.147 164.927 133.181C164.631 133.163 164.34 133.093 164.07 132.972C163.13 132.603 162.238 132.124 161.412 131.545C160.817 131.076 160.182 130.661 159.514 130.305C159.178 130.13 158.807 130.031 158.429 130.015C158.05 130 157.673 130.069 157.324 130.216C156.154 130.811 155.932 132.469 154.844 133.2C154.094 133.706 153.117 133.655 152.212 133.687C151.307 133.718 150.282 133.943 149.836 134.728C149.519 135.272 149.595 135.92 149.656 136.556L149.674 136.771C149.674 136.847 149.674 136.927 149.693 137.002C149.743 137.816 149.622 138.631 149.336 139.394C148.899 140.552 148.07 141.694 148.099 142.874C148.102 143.062 148.126 143.249 148.172 143.431C148.196 143.522 148.226 143.612 148.26 143.7C148.435 144.091 148.63 144.472 148.846 144.842L148.874 144.893C148.896 144.937 148.918 144.981 148.937 145.026C148.956 145.07 148.994 145.146 149.016 145.206C149.039 145.266 149.042 145.266 149.051 145.295C149.061 145.323 149.092 145.412 149.108 145.472C149.16 145.65 149.183 145.836 149.175 146.022C149.137 146.262 149.117 146.504 149.115 146.747C149.116 146.868 149.158 146.986 149.235 147.079C149.313 147.173 149.419 147.238 149.538 147.262C149.669 147.259 149.796 147.22 149.905 147.149C150.189 146.987 150.459 146.802 150.712 146.595C150.712 146.696 150.712 146.797 150.731 146.911C150.951 149.176 151.971 151.288 153.607 152.869C153.607 152.967 153.607 153.068 153.585 153.185C153.563 153.302 153.569 153.413 153.563 153.53L150.019 154.643C150.019 154.643 150.038 154.707 150.076 154.814L149.703 154.963C147.865 155.706 145.818 156.63 144.673 157.469C144.166 157.836 143.705 158.136 143.29 158.386C141.785 159.307 140.632 160.705 140.012 162.357L139.965 162.477C139.649 163.201 138.098 168.13 138.098 168.13C138.098 168.13 136.963 170.877 136.798 172.655C136.798 172.655 134.931 178.549 135.419 182.427C135.419 182.427 133.957 188.002 134.527 188.888C134.527 188.888 133.944 191.286 136.501 193.7C136.314 193.814 136.112 193.937 135.9 194.073C133.919 195.32 131.062 197.341 130.575 198.901C129.819 201.328 129.493 201.748 129.493 201.748L128.727 240.071C128.727 240.071 129.749 240.267 130.942 240.514L130.967 240.83L131.091 242.323L131.119 242.64C130.616 242.617 129.958 242.791 129.569 243.728C128.876 245.383 127.354 248.973 127.354 248.973C127.354 248.973 123.947 252.137 126.013 255.073C126.104 255.207 126.21 255.33 126.329 255.44C127.813 256.788 130.977 255.924 133.226 255.045C134.369 254.557 135.355 253.762 136.074 252.748C136.164 252.629 136.279 252.53 136.409 252.458C136.54 252.385 136.684 252.341 136.833 252.327C137.349 252.244 137.835 252.032 138.247 251.71C138.495 251.529 138.695 251.29 138.831 251.015C138.967 250.739 139.033 250.435 139.025 250.128C139.024 250.112 139.024 250.097 139.025 250.081L138.826 245.579C138.826 245.579 138.582 242.434 137.807 242.636L137.608 242.684L137.554 242.368L137.45 241.757L137.396 241.441H137.826C137.826 241.441 137.934 237.885 139.234 235.088C140.534 232.291 141.389 209.699 141.389 209.699C141.389 209.699 157.483 209.961 158.02 208.452C158.337 208.452 158.682 208.452 159.011 208.43C159.552 209.939 174.741 210.012 174.741 210.012C174.741 210.012 175.604 232.63 176.902 235.429C177.175 236.029 177.403 236.649 177.582 237.283C177.99 238.751 178.235 240.258 178.313 241.779H178.743L178.686 242.095L178.582 242.703L178.528 243.019L178.322 242.972C177.553 242.773 177.31 245.917 177.31 245.917L177.104 250.448C177.088 250.753 177.149 251.058 177.28 251.334C177.411 251.61 177.61 251.849 177.857 252.03C178.276 252.358 178.771 252.576 179.297 252.662C179.445 252.674 179.59 252.718 179.721 252.79C179.852 252.862 179.966 252.961 180.056 253.08C180.775 254.093 181.761 254.888 182.903 255.377C185.333 256.326 188.826 257.253 190.123 255.408C192.205 252.466 188.797 249.315 188.797 249.315ZM147.89 192.893H147.928H147.89ZM173.311 179.96C173.311 179.96 173.719 181.494 174.447 182.301C175.174 183.108 174.814 188.119 174.814 188.119C174.527 188.137 174.243 188.183 173.966 188.255C173.406 184.924 172.89 181.225 173.311 179.96Z" fill="url(#paint0_linear)"></path>
+                <path id="Vector_52" d="M138.317 243.111L131.955 243.744L131.777 241.615L131.654 240.15L131.48 238.008L137.491 238.324L137.969 241.074L138.07 241.672H138.073L138.317 243.111Z" fill="#8E6F7F"></path>
+                <path id="Vector_53" d="M184.65 238.328L184.476 240.466L184.355 241.931L184.175 244.06L177.81 243.427L178.063 241.988L178.161 241.393L178.638 238.644L184.65 238.328Z" fill="#8E6F7F"></path>
+                <path id="Vector_54" d="M177.908 207.516L158.989 208.405C158.128 208.447 157.267 208.447 156.407 208.405L138.333 207.519V193.2H177.908V207.516Z" fill="#473729"></path>
+                <path id="Vector_55" opacity="0.1" d="M177.908 207.516L158.989 208.405C158.128 208.447 157.267 208.447 156.407 208.405L138.333 207.519V193.2H177.908V207.516Z" fill="black"></path>
+                <path id="Vector_56" d="M153.68 154.621C153.68 154.621 147.378 156.738 145.005 158.478C144.511 158.839 144.056 159.136 143.657 159.38C142.192 160.283 141.07 161.651 140.471 163.265C140.455 163.309 140.44 163.347 140.424 163.382C140.107 164.094 138.605 168.918 138.605 168.918C138.605 168.918 137.497 171.607 137.339 173.347C137.181 175.088 146.356 179.754 146.356 179.754C146.356 179.754 141.927 189.403 143.904 191.855C145.881 194.307 174.75 195.494 174.118 191.855C173.485 188.217 172.378 182.206 172.931 180.545L182.739 173.427C182.739 173.427 182.343 163.461 179.259 160.851C176.174 158.241 172.298 157.292 172.298 157.292L166.158 155.71C166.158 155.71 163.994 161.088 161.621 161.8C159.248 162.512 154.05 159.225 153.68 154.621Z" fill="#DAE8F7"></path>
+                <g id="tie-shadow">
+                    <path id="Vector_57" opacity="0.1" d="M162.532 163.679C162.532 163.679 159.855 165.913 158.451 163.679C158.451 163.679 157.185 167.457 158.451 168.722L158.403 168.978L158.179 170.105L155.249 194.465L157.66 198.626L162.048 193.121L161.694 170.342L161.535 169.035L161.497 168.719H162.247C162.247 168.719 164.237 165.28 162.532 163.679Z" fill="black"></path>
+                    <path id="Vector_58" opacity="0.1" d="M163.165 163.679C163.165 163.679 160.488 165.913 159.084 163.679C159.084 163.679 157.818 167.457 159.084 168.722L159.036 168.978L158.811 170.105L155.882 194.465L158.293 198.626L162.681 193.121L162.326 170.342L162.168 169.035L162.13 168.719H162.88C162.88 168.719 164.87 165.28 163.165 163.679Z" fill="black"></path>
+                </g>
+                <path id="tie" d="M162.364 191.808L157.976 197.303L155.559 193.165L158.492 168.791L158.716 167.665L158.767 167.406L159.163 166.381L161.814 167.406L161.852 167.722L162.01 169.029L162.364 191.808Z" fill="#bb86fc"></path>
+                <path id="Vector_59" opacity="0.1" d="M137.969 241.07C136.604 241.062 135.242 240.923 133.903 240.656C133.242 240.491 132.413 240.308 131.657 240.147L131.48 238.005L137.491 238.321L137.969 241.07Z" fill="black"></path>
+                <path id="Vector_60" d="M158.024 208.433C157.499 209.911 141.813 209.651 141.813 209.651C141.813 209.651 140.971 231.798 139.706 234.54C138.44 237.283 138.333 240.76 138.333 240.76C138.333 240.76 135.593 240.76 133.903 240.34C132.214 239.919 129.474 239.39 129.474 239.39L130.214 201.821C130.214 201.821 130.531 201.426 131.268 199.05C131.743 197.522 134.527 195.541 136.453 194.304C137.529 193.627 138.333 193.184 138.333 193.184L147.691 193.564L153.468 193.798C153.468 193.798 158.552 206.956 158.024 208.433Z" fill="#17354C"></path>
+                <path id="Vector_61" opacity="0.1" d="M138.32 243.108L131.955 243.741L131.774 241.611C132.013 241.618 132.248 241.67 132.467 241.763C133.628 242.598 136.947 241.934 138.07 241.668L138.32 243.108Z" fill="black"></path>
+                <path id="Vector_62" d="M139.522 249.261C139.53 249.562 139.464 249.859 139.331 250.129C139.199 250.398 139.002 250.631 138.76 250.808C138.359 251.123 137.886 251.33 137.383 251.413C137.238 251.425 137.098 251.468 136.971 251.539C136.844 251.61 136.733 251.707 136.646 251.824C135.943 252.816 134.98 253.594 133.862 254.073C131.67 254.934 128.585 255.779 127.139 254.459C127.02 254.353 126.914 254.233 126.823 254.102C124.807 251.217 128.129 248.129 128.129 248.129C128.129 248.129 129.629 244.611 130.303 242.991C130.376 242.781 130.492 242.588 130.644 242.426C130.796 242.264 130.981 242.136 131.186 242.05C131.391 241.964 131.612 241.922 131.834 241.927C132.057 241.932 132.276 241.984 132.476 242.08C133.742 242.991 137.58 242.118 138.329 241.925C139.079 241.732 139.32 244.807 139.32 244.807L139.522 249.214C139.52 249.23 139.52 249.246 139.522 249.261Z" fill="#BCBCBD"></path>
+                <path id="Vector_63" opacity="0.1" d="M184.653 238.324L184.476 240.463C183.72 240.624 182.894 240.808 182.226 240.972C180.888 241.24 179.528 241.38 178.164 241.39L178.642 238.641L184.653 238.324Z" fill="black"></path>
+                <path id="Vector_64" d="M186.662 239.707C186.662 239.707 183.922 240.235 182.233 240.656C180.543 241.077 177.803 241.077 177.803 241.077C177.803 241.077 177.699 237.597 176.433 234.857C175.168 232.117 174.323 209.968 174.323 209.968C174.323 209.968 159.523 209.882 158.995 208.405C158.467 206.927 165.446 194.168 165.446 194.168L169.701 193.947L177.803 193.516C177.803 193.516 179.385 194.377 181.043 195.538C182.701 196.699 184.501 198.173 184.871 199.366C185.608 201.742 185.925 202.137 185.925 202.137L186.662 239.707Z" fill="#17354C"></path>
+                <path id="Vector_65" opacity="0.1" d="M184.355 241.928L184.178 244.057L177.81 243.424L178.063 241.985C179.186 242.254 182.492 242.915 183.662 242.08C183.882 241.986 184.117 241.934 184.355 241.928Z" fill="black"></path>
+                <path id="Vector_66" d="M183.663 242.396C183.864 242.3 184.083 242.248 184.305 242.242C184.528 242.237 184.749 242.278 184.955 242.364C185.16 242.45 185.345 242.579 185.498 242.741C185.65 242.904 185.767 243.097 185.839 243.307C186.51 244.927 188.013 248.448 188.013 248.448C188.013 248.448 191.335 251.533 189.319 254.418C188.054 256.228 184.65 255.32 182.283 254.39C181.165 253.912 180.201 253.134 179.499 252.14C179.412 252.023 179.301 251.926 179.173 251.855C179.046 251.784 178.904 251.741 178.759 251.729C178.245 251.64 177.762 251.422 177.354 251.096C177.115 250.919 176.923 250.685 176.795 250.416C176.668 250.147 176.609 249.85 176.623 249.552L176.823 245.123C176.823 245.123 177.06 242.045 177.81 242.238C178.56 242.431 182.397 243.307 183.663 242.396Z" fill="#BCBCBD"></path>
+                <path id="Vector_67" d="M167.309 157.292L160.922 164.062L160.744 164.252L160.513 164.116C159.178 163.325 152.5 159.294 153.072 158.082C153.389 157.393 153.575 155.956 153.673 154.719C153.673 154.571 153.692 154.425 153.705 154.283C153.768 153.305 153.784 152.546 153.784 152.546C153.784 152.546 165.806 150.489 165.332 152.546C165.139 153.378 165.427 154.314 165.835 155.134C165.955 155.371 166.088 155.596 166.218 155.808C166.537 156.334 166.902 156.83 167.309 157.292Z" fill="#FFD2C6"></path>
+                <path id="Vector_68" opacity="0.1" d="M165.835 155.134C164.041 156.456 161.836 157.096 159.613 156.94C157.391 156.784 155.296 155.842 153.705 154.283C153.768 153.305 153.784 152.546 153.784 152.546C153.784 152.546 165.806 150.489 165.332 152.546C165.148 153.384 165.427 154.314 165.835 155.134Z" fill="black"></path>
+                <path id="Vector_69" d="M169.654 147.259C169.688 149.694 168.775 152.048 167.107 153.822C165.439 155.597 163.147 156.654 160.714 156.769C158.281 156.885 155.899 156.051 154.07 154.443C152.241 152.835 151.108 150.579 150.912 148.151C150.886 147.856 150.874 147.559 150.874 147.259C150.851 146.012 151.077 144.773 151.538 143.615C151.999 142.456 152.687 141.401 153.561 140.511C154.435 139.621 155.477 138.914 156.627 138.431C157.777 137.949 159.012 137.7 160.259 137.7C161.506 137.7 162.741 137.949 163.891 138.431C165.041 138.914 166.083 139.621 166.957 140.511C167.831 141.401 168.518 142.456 168.98 143.615C169.441 144.773 169.667 146.012 169.644 147.259H169.654Z" fill="#FFD2C6"></path>
+                <path id="Vector_70" opacity="0.1" d="M169.536 145.76C169.292 145.959 169.009 146.104 168.705 146.186C168.4 146.269 168.082 146.286 167.771 146.237C166.322 145.921 165.825 143.706 164.345 143.555C162.864 143.403 161.88 145.453 160.381 145.412C159.748 145.396 159.178 144.981 158.596 144.693C157.577 144.19 155.901 143.884 154.8 144.263C153.145 144.83 152.585 146.361 151.528 147.579C151.348 147.786 151.15 147.978 150.937 148.151C150.911 147.856 150.899 147.559 150.899 147.259C150.877 144.885 151.756 142.591 153.358 140.839C154.961 139.087 157.167 138.008 159.534 137.817C161.9 137.627 164.251 138.341 166.112 139.815C167.973 141.288 169.207 143.413 169.565 145.76H169.536Z" fill="black"></path>
+                <path id="Vector_71" d="M158.596 144.38C159.178 144.665 159.735 145.083 160.381 145.098C161.877 145.136 162.855 143.077 164.342 143.238C165.829 143.4 166.319 145.595 167.774 145.921C168.752 146.143 169.673 145.412 170.403 144.731C170.876 144.355 171.261 143.882 171.533 143.343C171.849 142.558 171.685 141.653 171.39 140.856C171.096 140.059 170.682 139.293 170.536 138.455C170.359 137.449 170.565 136.348 170.087 135.443C169.889 135.101 169.618 134.807 169.293 134.583C168.968 134.359 168.598 134.209 168.208 134.146C167.425 134.024 166.624 134.115 165.889 134.408C165.53 134.582 165.143 134.689 164.747 134.725C164.457 134.707 164.173 134.637 163.908 134.519C162.993 134.158 162.124 133.69 161.32 133.124C160.744 132.664 160.126 132.257 159.476 131.909C159.149 131.737 158.788 131.641 158.419 131.626C158.05 131.611 157.683 131.679 157.344 131.823C156.205 132.406 155.986 134.038 154.926 134.744C154.196 135.237 153.243 135.186 152.364 135.221C151.484 135.256 150.466 135.468 150.048 136.237C149.675 136.898 149.861 137.711 149.909 138.467C149.956 139.264 149.836 140.061 149.558 140.808C149.067 142.118 148.071 143.409 148.425 144.763C148.656 145.652 149.453 146.389 149.406 147.294C149.368 147.527 149.348 147.763 149.346 148C149.347 148.119 149.388 148.234 149.464 148.326C149.539 148.418 149.644 148.481 149.76 148.506C149.887 148.504 150.01 148.465 150.114 148.392C150.64 148.098 151.112 147.716 151.51 147.262C152.579 146.041 153.133 144.513 154.787 143.947C155.904 143.577 157.578 143.874 158.596 144.38Z" fill="#585268"></path>
+                <path id="Vector_72" d="M169.483 199.607L167.774 200.072L163.829 201.157C162.916 201.837 161.888 202.348 160.795 202.666C158.381 203.276 156.821 201.948 155.524 199.891C155.002 199.069 155.265 198.411 155.948 197.895C157.593 196.652 161.662 196.247 162.908 196.411C163.727 196.478 164.55 196.35 165.31 196.038C165.849 195.842 166.369 195.596 166.863 195.304L169.483 199.607Z" fill="#FFD2C6"></path>
+                <path id="Vector_73" opacity="0.1" d="M169.483 199.606L167.774 200.072C167.284 198.594 166.477 196.74 165.322 196.028C165.862 195.832 166.382 195.586 166.876 195.294L169.483 199.606Z" fill="black"></path>
+                <path id="Vector_74" opacity="0.1" d="M181.037 195.538C180.328 196.111 179.727 196.633 179.727 196.633C179.727 196.633 171.501 199.211 169.919 202.454C169.919 202.454 168.654 196.126 165.965 196.047C165.965 196.047 168.859 194.912 169.695 193.94L177.803 193.516C177.803 193.516 179.385 194.377 181.037 195.538Z" fill="black"></path>
+                <path id="Vector_75" d="M180.913 170.737L182.732 173.427L183.286 181.81C183.286 181.81 183.444 189.878 183.286 190.194L183.444 189.878C183.444 189.878 183.128 193.042 182.179 193.595C181.23 194.149 179.094 196 179.094 196C179.094 196 170.868 198.578 169.286 201.821C169.286 201.821 168.021 195.494 165.332 195.415C165.332 195.415 169.366 193.833 169.286 192.804C169.286 192.804 172.371 190.036 172.608 189.324C172.846 188.612 174.386 188.533 174.386 188.533C174.386 188.533 174.744 183.63 174.032 182.839C173.32 182.048 172.925 180.545 172.925 180.545L180.913 170.737Z" fill="#DAE8F7"></path>
+                <path id="Vector_76" opacity="0.1" d="M160.795 202.656C158.381 203.267 156.821 201.938 155.524 199.882C155.002 199.059 155.265 198.401 155.948 197.885L156.195 197.983C156.195 197.983 160.545 199.869 160.795 202.656Z" fill="black"></path>
+                <path id="Vector_77" d="M159.795 205.184C156.948 209.139 151.649 203.64 151.253 202.533C151.111 202.131 150.57 201.774 149.947 201.486C149.156 201.14 148.335 200.868 147.495 200.673L152.202 196.838L153.104 197.199L155.879 198.309C155.879 198.309 162.643 201.229 159.795 205.184Z" fill="#FFD2C6"></path>
+                <path id="Vector_78" opacity="0.1" d="M153.104 197.199C150.51 198.433 149.956 200.208 149.94 201.486C149.149 201.14 148.329 200.868 147.488 200.673L152.196 196.838L153.104 197.199Z" fill="black"></path>
+                <path id="Vector_79" opacity="0.1" d="M148.96 204.036C148.96 204.036 141.367 198.025 137.728 195.415C137.273 195.09 136.849 194.724 136.463 194.32C137.539 193.643 138.342 193.2 138.342 193.2L147.7 193.58C147.95 194.013 148.099 194.307 148.099 194.307L153.161 197.629C147.061 199.607 148.96 204.036 148.96 204.036Z" fill="black"></path>
+                <path id="Vector_80" d="M138.835 172.477L137.333 173.347C137.333 173.347 135.514 179.121 135.988 182.918C135.988 182.918 134.564 188.375 135.118 189.245C135.118 189.245 134.406 192.172 138.045 194.782C141.683 197.392 149.276 203.403 149.276 203.403C149.276 203.403 147.378 198.974 153.468 196.996L148.406 193.674C148.406 193.674 146.707 190.302 145.084 189.562C145.084 189.562 143.818 188.296 144.609 187.426C145.4 186.556 146.033 185.607 146.033 184.737C146.033 183.867 146.349 179.754 146.349 179.754L147.298 175.325L138.835 172.477Z" fill="#DAE8F7"></path>
+                <path id="Vector_81" opacity="0.1" d="M161.852 168.719H158.767C158.749 168.702 158.732 168.682 158.717 168.662L158.767 168.403L159.163 167.378L161.814 168.403L161.852 168.719Z" fill="black"></path>
+                <path id="Vector_82" d="M162.848 163.363C162.848 163.363 160.172 165.596 158.767 163.363C158.767 163.363 157.502 167.14 158.767 168.406H162.564C162.564 168.406 164.553 164.964 162.848 163.363Z" fill="#bb86fc"></path>
+                <path id="Vector_83" opacity="0.1" d="M166.468 156.026L170.454 157.054C170.454 157.054 166.676 166.862 165.844 166.862L161.061 164.568L166.468 156.026Z" fill="black"></path>
+                <path id="Vector_84" opacity="0.1" d="M167.309 157.291L160.922 164.062L160.744 164.252L160.513 164.116L166.097 155.599L166.233 155.824C166.549 156.343 166.909 156.834 167.309 157.291Z" fill="black"></path>
+                <path id="Vector_85" d="M166.151 155.71L170.138 156.738C170.138 156.738 166.36 166.545 165.528 166.545L160.744 164.252L166.151 155.71Z" fill="#DAE8F7"></path>
+                <path id="Vector_86" opacity="0.1" d="M153.363 154.938L149.909 156.026C149.909 156.026 153.31 166.308 154.259 166.862L160.428 164.568L153.363 154.938Z" fill="black"></path>
+                <path id="Vector_87" opacity="0.1" d="M160.921 164.062L160.744 164.252C160.744 164.252 152.439 159.427 153.072 158.082C153.389 157.393 153.575 155.956 153.673 154.719V154.526L160.921 164.062Z" fill="black"></path>
+                <path id="Vector_88" d="M153.68 154.621L150.225 155.71C150.225 155.71 153.626 165.992 154.575 166.545L160.745 164.252L153.68 154.621Z" fill="#DAE8F7"></path>
+                <path id="Vector_89" opacity="0.1" d="M145.451 171.418C145.451 171.418 147.533 171.788 147.495 173.949C147.457 176.109 148.352 171.155 148.352 171.155L145.451 171.418Z" fill="black"></path>
+                <path id="Vector_90" opacity="0.1" d="M145.926 178.061C145.926 178.061 145.347 181.33 145.926 184.285C146.504 187.239 146.454 188.239 146.903 188.397C147.352 188.556 145.926 178.061 145.926 178.061Z" fill="black"></path>
+                <path id="Vector_91" opacity="0.1" d="M172.501 177.745C172.335 177.766 172.175 177.823 172.034 177.913C171.892 178.002 171.772 178.121 171.681 178.262C171.591 178.403 171.532 178.562 171.51 178.728C171.488 178.894 171.502 179.063 171.552 179.223C171.818 180.545 172.501 177.745 172.501 177.745Z" fill="black"></path>
+                <path id="Vector_92" opacity="0.1" d="M136.33 188.818C136.33 188.818 143.11 188.239 144.239 189.451C145.369 190.663 136.33 188.818 136.33 188.818Z" fill="black"></path>
+                <path id="Vector_93" opacity="0.1" d="M175.665 189.277C175.665 189.277 180.622 188.666 181.783 189.511L175.665 189.277Z" fill="black"></path>
+                <path id="Vector_94" opacity="0.1" d="M177.67 190.084C177.67 190.084 180.622 190.188 181.309 190.558L177.67 190.084Z" fill="black">
+                </path>
+                <path id="Vector_95" opacity="0.1" d="M178.249 191.144C178.249 191.144 180.938 191.039 181.309 191.46L178.249 191.144Z" fill="black"></path>
+                <path id="Vector_96" opacity="0.1" d="M140.326 206.013C140.326 206.013 142.629 207.212 142.395 208.436C142.161 209.661 140.326 206.013 140.326 206.013Z" fill="black"></path>
+                <path id="Vector_97" opacity="0.1" d="M138.333 206.595C138.333 206.595 141.847 211.154 141.37 212.657C140.892 214.16 138.333 206.595 138.333 206.595Z" fill="black"></path>
+                <path id="Vector_98" opacity="0.1" d="M176.009 207.351C176.009 207.351 174.269 208.784 174.317 209.968L176.009 207.351Z" fill="black"></path>
+                <path id="Vector_99" opacity="0.1" d="M135.83 235.594C135.83 235.594 135.83 236.939 138.32 238.916L135.83 235.594Z" fill="black">
+                </path>
+                <path id="Vector_100" opacity="0.1" d="M179.778 232.43C179.778 232.43 179.512 235.227 177.088 236.676L179.778 232.43Z" fill="black">
+                </path>
+                <g id="Group" opacity="0.1">
+                    <path id="Vector_101" opacity="0.1" d="M149.909 137.506C149.909 137.287 149.871 137.063 149.852 136.841C149.811 137.324 149.826 137.81 149.896 138.29C149.918 138.029 149.922 137.767 149.909 137.506Z" fill="black"></path>
+                    <path id="Vector_102" opacity="0.1" d="M149.406 146.345C149.453 145.427 148.656 144.69 148.425 143.814C148.413 143.771 148.405 143.728 148.399 143.684C148.327 144.041 148.336 144.41 148.425 144.763C148.611 145.488 149.175 146.108 149.352 146.81C149.368 146.646 149.399 146.497 149.406 146.345Z" fill="black"></path>
+                    <path id="Vector_103" opacity="0.1" d="M171.533 142.393C171.26 142.932 170.875 143.406 170.403 143.782C169.666 144.463 168.752 145.193 167.774 144.972C166.322 144.656 165.822 142.441 164.342 142.289C162.861 142.137 161.877 144.187 160.381 144.146C159.748 144.13 159.178 143.716 158.596 143.428C157.578 142.925 155.904 142.618 154.8 142.998C153.145 143.564 152.585 145.095 151.522 146.313C151.125 146.767 150.655 147.15 150.13 147.446C150.024 147.516 149.9 147.554 149.773 147.557C149.693 147.546 149.618 147.514 149.556 147.464C149.493 147.413 149.446 147.347 149.418 147.272V147.294C149.381 147.529 149.361 147.765 149.358 148.003C149.36 148.121 149.402 148.236 149.477 148.327C149.553 148.418 149.657 148.481 149.773 148.506C149.9 148.504 150.024 148.465 150.13 148.395C150.655 148.099 151.125 147.716 151.522 147.262C152.591 146.044 153.145 144.513 154.8 143.947C155.917 143.567 157.59 143.874 158.596 144.377C159.178 144.665 159.735 145.079 160.381 145.095C161.877 145.133 162.855 143.074 164.342 143.238C165.829 143.403 166.322 145.592 167.774 145.921C168.752 146.142 169.672 145.412 170.403 144.731C170.875 144.355 171.26 143.881 171.533 143.343C171.709 142.885 171.756 142.388 171.669 141.906C171.642 142.073 171.596 142.237 171.533 142.393Z" fill="black"></path>
+                </g>
+                <path id="Vector_104" opacity="0.1" d="M128.133 248.132C128.133 248.132 130.717 247.335 131.79 248.847L128.133 248.132Z" fill="black"></path>
+                <path id="Vector_105" opacity="0.1" d="M188.013 248.448C188.013 248.448 183.732 248.198 183.9 249.356L188.013 248.448Z" fill="black"></path>
+            </g>
+        </svg>
+                    </div>
+                    <div class="dropdown px-md-5 col-md-6 d-flex flex-column align-items-center justify-content-center animate__animated animate__fadeInUp animate__delay-1s">
+                        <h2 class="main-heading mb-0 experience_head">My <span class="my_experience">Experience</span></h2>
+                        <p class="pre-heading font-weight-bolder text-center animate__animated animate__fadeInUp animate__delay-2s">
+                            Nothing ever becomes real 'til it is experienced.― John Keats
+                        </p>
+                    </div>
+                </div>
+            </div>
+        </section>
+    </div>
+    </div>
+
+    )
+    
+}
+
+export default topIlus;
